将日志输出函数的参数类型从int改为size_t，以增强类型安全性

+++ cpp/logging/async_logging.cc
@@ -65,7 +65,7 @@ namespace lsm {
 
     thread_local ThreadBuffer t_buffer;
 
-    static void asyncOutput(const char* msg, int len) {
+    static void asyncOutput(const char* msg, size_t len) {
         t_buffer.append(msg, len);
     }
