+++ cpp/logging/logging.cc
@@ -15,13 +15,13 @@ namespace lsm
     __thread char t_time[64];
     __thread time_t t_last_second;
 
-    void ConsoleOutput(const char *msg, int len)
+    void ConsoleOutput(const char *msg, size_t len)
     {
         size_t n = fwrite(msg, 1, len, stdout);
         (void)n;
     }
     
-    void defaultOutput(const char *msg, int len)
+    void defaultOutput(const char *msg, size_t len)
     {
         ConsoleOutput(msg, len);
     }
